backend/requirements.txt
aiortc==1.9.0
av==12.3.0
opencv-python==4.10.0.84
python-dotenv==1.0.1
websockets==12.0

